Finishing work
- Added more sounds for different balls
- Setup different balls
- Reorginize scripts
- Added rules to the start page

+++ Assets/Scripts/UI/Screens/TeamResultElement.cs
@@ -12,7 +12,7 @@ namespace Janegamedev.UI.Screens
         [SerializeField]
         private TextMeshProUGUI score;
 
-        public void DisplayTeamScore(int teamId, int teamScore)
+        public void DisplayTeamScore(int teamId, long teamScore)
         {
             teamName.text = string.Format(TEAM_NAME, teamId + 1);
             score.text = teamScore.ToString();
